refactor: feed files into feed
